Disable fully buffered mode.

+++ mona/core/monalibc/stdio/fileno.c
@@ -35,5 +35,5 @@
 
 int fileno(FILE *fp)
 {
-	return fp->_file;
+	return (int)fp->_file;
 }
